requirements.txt
asgiref==3.5.2
dj-database-url==1.0.0
Django==3.2.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-phonenumber-field==6.3.0
django-pwa==1.0.10
geographiclib==1.52
geopy==2.2.0
gunicorn==20.1.0
phonenumbers==8.12.52
Pillow==9.2.0
postgis==1.0.4
psycopg2==2.9.3
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
whitenoise==6.2.0

